backend/requirements.txt
django
prettyconf
djangorestframework
django-filter
django-qr-code
markdown
Faker #Libreria para generar datos random
django_otp
django-cors-headers
Pillow
qrcode
phonenumbers
rest_framework_simplejwt
